ENGA-1273 DRY code. Extract common function updateEventListeners to utils
